Remove redundant cast

+++ NightgamesMod/nightgames/characters/WillpowerStat.java
@@ -19,7 +19,7 @@ public class WillpowerStat extends Meter {
     }
 
     public void renew() {
-        current = (int) max;
+        current = max;
     }
 
     public void recover(int i) {
